Title: 작지만 큰 숫자, Time: 103ms, Memory: 1MB, Status: Memory Limit Exceed - CodeTree

+++ 241103/작지만 큰 숫자/small-but-big-number.cpp
@@ -24,8 +24,8 @@ string t;
 int main() {
 
     cin >> n >> m;
-    set<long long> s;
-    long long a, b;
+    set<int> s;
+    int a, b;
     for (int i = 0; i < n; i++) {
         cin >> a;
         s.insert(a);
